Separated medicine card and medicine page into components
